完成初赛

+++ os/src/syscall/mod.rs
@@ -88,7 +88,7 @@ pub fn syscall(syscall_id: usize, args: [usize; 6]) -> isize {
         SYSCALL_OPENAT => sys_openat(args[0],args[1] as *const u8,args[2] as u32,args[3] as u32,),
         SYSCALL_CLOSE => sys_close(args[0]),
         SYSCALL_GETDENTS64 => sys_getdents64(args[0], args[1] as *mut u8, args[2]),
-        SYSCALL_PIPE => sys_pipe(args[0] as *mut usize),
+        SYSCALL_PIPE => sys_pipe(args[0] as *mut u32),
         SYSCALL_MKDIRAT => sys_mkdirat(args[0], args[1] as *const u8, args[2] as u32),
         SYSCALL_UMOUNT2 => sys_umount2(args[0] as *const u8, args[1] as u32),
         SYSCALL_MOUNT => sys_mount(args[0] as *const u8,args[1] as *const u8,args[2] as *const u8,args[3], args[4] as *const u8,),
